+++ RELEASENOTES.md
@@ -179,6 +179,7 @@
     *   Add `MiniController` composable to `media3-ui-compose` module. It
         provides a compact affordance to control the `Player` while displaying
         the title, artist, artwork, and progress of the current media item.
+    *   Publish `PlayerDefaults` object with composable for `BottomControls`.
 *   Downloads:
 *   OkHttp extension:
 *   Cronet extension: